Mirrors branch main from private at 9c7cceefc844b0fa24e163690bf064b7c54a8435 (#15)
What

This PR updates the mirror with the latest in Boojum before FOSing the repo.

+++ src/cs/implementations/lookup_placement.rs
@@ -28,7 +28,7 @@ impl<
 
         let non_zero_table_id = table_id;
 
-        let mut principal_width = lookup_width as usize;
+        let mut principal_width = lookup_width;
         if id_in_constant == false {
             principal_width += 1;
         }
